requirements.txt
# ── Core Flask ───────────────────────────────────────────────
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.10.0
Flask-Limiter==3.8.0
Flask-SocketIO==5.3.7

# ── Database ─────────────────────────────────────────────────
PyMySQL==1.1.1             # MySQL driver
cryptography==43.0.1       # required by PyMySQL for auth
SQLAlchemy==2.0.36

# ── Background Jobs ──────────────────────────────────────────
celery==5.4.0
redis==5.1.1               # Redis client (also used by rate limiter)
kombu==5.3.4               # Celery transport

# ── Auth & Security ──────────────────────────────────────────
Werkzeug==3.0.4
itsdangerous==2.2.0
WTForms==3.1.2
bcrypt==4.2.0

# ── Image Processing ─────────────────────────────────────────
Pillow==11.0.0
opencv-python-headless==4.10.0.84   # headless — no GUI
face-recognition==1.3.0
pytesseract==0.3.13
numpy==2.1.3

# ── Real-time ────────────────────────────────────────────────
python-socketio==5.11.4
eventlet==0.37.0            # or use gevent
simple-websocket==1.1.0

# ── HTTP & Networking ────────────────────────────────────────
requests==2.32.3
python-dotenv==1.0.1

# ── Utilities ────────────────────────────────────────────────
python-dateutil==2.9.0

# ── Production Server ────────────────────────────────────────
gunicorn==23.0.0

# ── Monitoring (optional but recommended) ────────────────────
# sentry-sdk[flask]==2.19.0